requirements.txt
aiofiles==23.2.1
aiogram==3.13.0
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asyncio==3.4.3
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.1
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
idna==3.7
imutils==0.5.4
kiwisolver==1.4.7
magic-filter==1.0.12
matplotlib==3.9.2
multidict==6.0.5
numpy==1.26.4
opencv-python==4.9.0.80
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
prettytable==3.12.0
propcache==0.2.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4